Refactor folder structure in basicElements
