+++ src/components/Router/Router.tsx
@@ -2,7 +2,7 @@ import React, {ReactElement} from "react";
 import {Redirect, Route, Switch} from "react-router-dom";
 import Health from "../pages/Health";
 import Error404 from "../pages/errors/Error404";
-import Login from "../basicElements/Login";
+import Login from "../basicElements/login/Login";
 import Registration from "../pages/User/Registration";
 import FileSystem, {filesBaseUrl} from "../pages/filesytem/Filesystem";
 import Profile from "../pages/User/Profile";
