+++ src/components/basicElements/bottomArea/BottomArea.tsx
@@ -1,0 +1,13 @@
+import Footer from "./Footer";
+import { BottomBanner } from "./BottomBanner";
+import React from "react";
+
+
+export default function BottomArea() {
+    return(
+        <>
+            <Footer />
+            <BottomBanner/>
+        </>
+    )
+}
