+++ src/components/basicElements/bottomArea/BottomBanner.tsx
