+++ src/components/basicElements/bottomArea/Footer.tsx
