+++ src/components/basicElements/login/Login.test.tsx
@@ -1,7 +1,7 @@
 import { render } from "@testing-library/react";
 import React from "react";
 import Login from "./Login";
-import {RouterWrapper} from "../../dev/testUtils/RouterWrapper";
+import {RouterWrapper} from "../../../dev/testUtils/RouterWrapper";
 
 test("renders Login page", () => {
     const breadcrumb = render(
