+++ src/components/basicElements/topArea/TopArea.tsx
@@ -1,0 +1,13 @@
+import { TopBanner } from "./TopBanner";
+import Header from "./Header";
+import React from "react";
+
+
+export default function HeadArea(){
+    return(
+        <>
+            <TopBanner />
+            <Header />
+        </>
+    )
+}
