+++ src/components/basicElements/topArea/TopBanner.tsx
